requirements.txt
numpy
pandas
seaborn
matplotlib
scikit-learn
python-dotenv
google-generativeai
streamlit
supabase